Allow plotting tools to be imported, returning plots
