+++ nyx-py/nyx_space/plots/plot_od_rslt.py
@@ -1,37 +1,18 @@
 import click
-import numpy as np
-import plotly.express as px
-import plotly.graph_objects as go
 import polars as pl
-from plotly.subplots import make_subplots
-from scipy import stats
 
-from nyx_space.plots import TEMPLATE, convert_units, watermark
+from nyx_space.plots.od import (
+    cr_cd,
+    kalman_gains,
+    filter_smoother_ratios,
+    residuals,
+    uncertainty,
+    od_dashboard,
+    orbital_element_uncertainty,
+)
+from nyx_space.plots.md import ric_diff
 
 optional_est_params = ["cr", "cd"]
-
-
-def autocorr(x: np.ndarray, max_lag: int) -> np.ndarray:
-    x = np.asarray(x, dtype=float)
-    x = x[np.isfinite(x)]
-
-    if len(x) < 3:
-        return np.array([])
-
-    x = x - np.mean(x)
-    denom = np.dot(x, x)
-
-    if denom == 0.0:
-        return np.zeros(max_lag + 1)
-
-    max_lag = min(max_lag, len(x) - 1)
-
-    return np.array(
-        [
-            np.dot(x[:-lag], x[lag:]) / denom if lag > 0 else 1.0
-            for lag in range(max_lag + 1)
-        ]
-    )
 
 
 @click.command
@@ -41,424 +22,30 @@ def autocorr(x: np.ndarray, max_lag: int) -> np.ndarray:
 def main(path: str, wstats: bool, error_ric: str):
     df = pl.read_parquet(path)
 
-    df = df.with_columns(
-        pl.col("Epoch (UTC)").str.to_datetime("%Y-%m-%dT%H:%M:%S%.f")
-    ).sort("Epoch (UTC)", descending=False)
-    df = convert_units(df)
-
     if error_ric:
         ricdf = pl.read_parquet(error_ric)
-        ricdf = ricdf.with_columns(
-            pl.col("Epoch (UTC)").str.to_datetime("%Y-%m-%dT%H:%M:%S%.f")
-        ).sort("Epoch (UTC)", descending=False)
-        ricdf = convert_units(ricdf)
+        ric_diff(ricdf).show()
 
-        ric_fig = make_subplots(
-            rows=2,
-            cols=1,
-            subplot_titles=["RIC position error (m)", "RIC velocity error (m/s)"],
-            vertical_spacing=0.1,
-        )
-        this_ric_fig = px.line(
-            ricdf,
-            x="Epoch (UTC)",
-            y=["Delta X (RIC) (m)", "Delta Y (RIC) (m)", "Delta Z (RIC) (m)"],
-            template=TEMPLATE,
-        )
-        for trace in this_ric_fig.data:
-            ric_fig.add_trace(trace, row=1, col=1)
+    residuals(df, path).show()
+    uncertainty(df, 3.0, path).show()
+    for dash in od_dashboard(df, path):
+        dash.show()
 
-        this_ric_fig = px.line(
-            ricdf,
-            x="Epoch (UTC)",
-            y=[
-                "Delta Vx (RIC) (m/s)",
-                "Delta Vy (RIC) (m/s)",
-                "Delta Vz (RIC) (m/s)",
-            ],
-            template=TEMPLATE,
-        )
-        for trace in this_ric_fig.data:
-            ric_fig.add_trace(trace, row=2, col=1)
-        watermark(ric_fig, path).show()
+    cr_cd_plot = cr_cd(df, path)
+    if cr_cd_plot is not None:
+        cr_cd_plot.show()
 
-    all_msr_types = ["Range (m)", "Doppler (m/s)", "Azimuth (deg)", "Elevation (deg)"]
-    msr_type_count = 0
-    msr_types = []
-
-    for msr_type in all_msr_types:
-        if f"Measurement noise: {msr_type}" in df.columns:
-            print(f"Found data for {msr_type}")
-            msr_type_count += 1
-            msr_types += [msr_type]
-            # Add the +/- 3 sigmas on measurement noise
-            df = df.with_columns(
-                [
-                    (3.0 * pl.col(f"Measurement noise: {msr_type}")).alias(
-                        f"Measurement noise 3-Sigma: {msr_type}"
-                    ),
-                    (-3.0 * pl.col(f"Measurement noise: {msr_type}")).alias(
-                        f"Measurement noise -3-Sigma: {msr_type}"
-                    ),
-                ]
-            )
-
-    # Plot the measurement residuals and their noises.
-    fig = make_subplots(
-        rows=len(msr_types),
-        cols=1,
-        subplot_titles=[msr for msr in msr_types],
-        vertical_spacing=0.1,
-    )
-    legend_added = set()  # Track which trace names are already in legend
-
-    for idx, msr in enumerate(msr_types, start=1):
-        unit = msr.split()[-1][1:-1]
-        y_cols = [
-            f"{col}: {msr}"
-            for col in [
-                "Prefit residual",
-                "Postfit residual",
-                "Measurement noise 3-Sigma",
-                "Measurement noise -3-Sigma",
-            ]
-        ]
-        for y in y_cols[:2]:
-            fig.add_trace(
-                go.Scatter(
-                    x=df["Epoch (UTC)"],
-                    y=df[y],
-                    mode="markers",
-                    name=y,
-                    legendgroup=y,
-                    marker=dict(color="blue" if "Prefit" in y else "red"),
-                    showlegend=True,
-                ),
-                row=idx,
-                col=1,
-            )
-
-        # Add 3-sigma bounds
-        for y in y_cols[2:]:
-            trace_type = "3-Sigma bounds"
-            fig.add_trace(
-                go.Scatter(
-                    x=df["Epoch (UTC)"],
-                    y=df[y],
-                    mode="lines",
-                    name=trace_type,
-                    line=dict(color="black"),
-                    legendgroup=trace_type,
-                    connectgaps=True,
-                    showlegend=(trace_type not in legend_added),
-                ),
-                row=idx,
-                col=1,
-            )
-            legend_added.add(trace_type)
-
-        fig.update_yaxes(title_text=unit, row=idx, col=1)
-    fig.update_layout(title_text="Measurement Residuals", template=TEMPLATE)
-    fig.update_xaxes(matches="x")
-    watermark(fig, path).show()
-
-    # Plot the RIC uncertainty for position and velocity
-    sigma_fig = make_subplots(
-        rows=2,
-        cols=1,
-        subplot_titles=[
-            "RIC 1-sigma position uncertainty (m)",
-            "RIC 1-sigma velocity uncertainty (m/s)",
-        ],
-        vertical_spacing=0.1,
-    )
-    this_fig = px.line(
-        df,
-        x="Epoch (UTC)",
-        y=["Sigma X (RIC) (m)", "Sigma Y (RIC) (m)", "Sigma Z (RIC) (m)"],
-        template=TEMPLATE,
-    )
-    for trace in this_fig.data:
-        sigma_fig.add_trace(trace, row=1, col=1)
-
-    this_fig = px.line(
-        df,
-        x="Epoch (UTC)",
-        y=["Sigma Vx (RIC) (m/s)", "Sigma Vy (RIC) (m/s)", "Sigma Vz (RIC) (m/s)"],
-        template=TEMPLATE,
-    )
-    for trace in this_fig.data:
-        sigma_fig.add_trace(trace, row=2, col=1)
-    watermark(sigma_fig, path).show()
-
-    # Create one QQ plot per whitened residual (there is only one if scalar processing of measurements)
-    whitened_resids = [c for c in df.columns if "Whitened residual" in c]
-    if len(whitened_resids) == len(msr_types):
-        # Vectorized filter, so the whitened residuals map directly the measurement types, in that order
-        resid_titles = msr_types
-    else:
-        resid_titles = whitened_resids
-
-    # Plot the residual ratios
-    for resid_title, whitened_resid in zip(resid_titles, whitened_resids):
-        # 1. Create the subplot figure: 2 rows, 1 column
-        # shared_xaxes=True links zooming/panning across both rows
-        fig = make_subplots(
-            rows=4,
-            cols=2,
-            shared_xaxes=False,
-            vertical_spacing=0.08,
-            horizontal_spacing=0.07,
-            specs=[
-                [{"colspan": 2}, None],
-                [{"colspan": 2}, None],
-                [{"colspan": 2}, None],
-                [{}, {}],
-            ],
-            subplot_titles=(
-                "Rejected Status",
-                "By Tracker",
-                "Autocorrelation by Tracker",
-                "Accepted Residuals Histogram",
-                "Accepted Residuals QQ Plot",
-            ),
-        )
-        # Generate the "Rejected" plot
-        # We use px to get the traces, then add them to the subplot
-        fig_rej = px.scatter(
-            df, x="Epoch (UTC)", y=whitened_resid, color="Residual Rejected"
-        )
-        for trace in fig_rej.data:
-            fig.add_trace(trace, row=1, col=1)
-
-        # Generate the "Tracker" plot
-        fig_track = px.scatter(df, x="Epoch (UTC)", y=whitened_resid, color="Tracker")
-        for trace in fig_track.data:
-            # We set showlegend=True to avoid duplicate legend entries if needed
-            fig.add_trace(trace, row=2, col=1)
-
-        df_accepted = df.filter(~pl.col("Residual Rejected"))
-        fig_hist = px.histogram(
-            df_accepted, x=whitened_resid, color="Tracker", barmode="overlay"
-        )
-        for trace in fig_hist.data:
-            trace.showlegend = False
-            fig.add_trace(trace, row=4, col=1)
-
-        # Generate the QQ plot
-        sample = df_accepted[whitened_resid].drop_nulls().to_numpy()
-
-        # QQ data
-        (osm, osr), (slope, intercept, r) = stats.probplot(
-            sample, dist="norm", fit=True
-        )
-
-        x_line = np.linspace(np.min(osm), np.max(osm), 200)
-
-        # QQ scatter
-        fig.add_trace(
-            go.Scatter(
-                x=osm,
-                y=osr,
-                mode="markers",
-                name=f"{whitened_resid} QQ",
-                marker=dict(color="blue"),
-            ),
-            row=4,
-            col=2,
-        )
-
-        # Standard-normal expected line: this is the important one for whitened residuals
-        fig.add_trace(
-            go.Scatter(
-                x=x_line,
-                y=x_line,
-                mode="lines",
-                name="Expected N(0,1)",
-                line=dict(color="red", dash="dash"),
-            ),
-            row=4,
-            col=2,
-        )
-
-        # Optional fitted line, useful to see empirical bias/scale
-        fig.add_trace(
-            go.Scatter(
-                x=x_line,
-                y=slope * x_line + intercept,
-                mode="lines",
-                name=f"Fitted normal: μ={intercept:.3f}, σ={slope:.3f}, R={r:.3f}",
-                line=dict(color="gray"),
-            ),
-            row=4,
-            col=2,
-        )
-
-        for tracker in df_accepted["Tracker"].unique().to_list():
-            tracker_df = df_accepted.filter(pl.col("Tracker") == tracker).sort(
-                "Epoch (UTC)"
-            )
-
-            x = tracker_df[whitened_resid].drop_nulls().to_numpy()
-
-            if len(x) < 5:
-                continue
-
-            max_lag = 30
-            rho = autocorr(x, max_lag=max_lag)
-            lags = np.arange(len(rho))
-
-            fig.add_trace(
-                go.Bar(
-                    x=lags,
-                    y=rho,
-                    name=f"{tracker} ACF",
-                    showlegend=False,
-                    opacity=0.7,
-                ),
-                row=3,
-                col=1,
-            )
-
-            # Approximate 95% white-noise bounds
-            bound = 1.96 / np.sqrt(len(x))
-
-            fig.add_trace(
-                go.Scatter(
-                    x=[0, len(rho) - 1],
-                    y=[bound, bound],
-                    mode="lines",
-                    line=dict(dash="dash"),
-                    name=f"{tracker} +95%",
-                    showlegend=False,
-                ),
-                row=3,
-                col=1,
-            )
-
-            fig.add_trace(
-                go.Scatter(
-                    x=[0, len(rho) - 1],
-                    y=[-bound, -bound],
-                    mode="lines",
-                    line=dict(dash="dash"),
-                    name=f"{tracker} -95%",
-                    showlegend=False,
-                ),
-                row=3,
-                col=1,
-            )
-
-        # Global Layout Updates
-        fig.update_layout(
-            title_text=f"Residual Analysis: {resid_title}",
-            template=TEMPLATE,
-        )
-        # Update axes titles for clarity
-        fig.update_yaxes(title_text=whitened_resid, row=1, col=1)
-        fig.update_yaxes(title_text=whitened_resid, row=2, col=1)
-        fig.update_yaxes(title_text="Count", row=3, col=1)
-        fig.update_xaxes(title_text="Theoretical N(0,1) Quantiles", row=3, col=2)
-        fig.update_yaxes(title_text="Sample Quantiles", row=3, col=2)
-
-        watermark(fig, path).show()
-
-    # Plot the optional parameters that were estimated.
-    plots_to_make = []
-    for col in optional_est_params:
-        if df[col].max() != df[col].min():
-            # We should plot this!
-            sigma_col = [
-                c for c in df.columns if col in c.lower() and "sigma" in c.lower()
-            ][0]
-            plots_to_make += [(col.capitalize(), col, sigma_col)]
-
-    if not plots_to_make:
-        print(
-            f"Neither of {optional_est_params} were estimated in this run, skipping plots"
-        )
-    else:
-        fig = make_subplots(
-            rows=len(plots_to_make),
-            cols=1,
-            subplot_titles=[p[0] for p in plots_to_make],
-            vertical_spacing=0.1,
-        )
-
-        legend_added = False
-
-        for idx, (title, val_col, sigma_col) in enumerate(plots_to_make, start=1):
-            # Add the estimate trace
-            fig.add_trace(
-                go.Scatter(
-                    x=df["Epoch (UTC)"],
-                    y=df[val_col],
-                    mode="lines+markers",
-                    name=title,
-                    legendgroup=title,
-                    marker=dict(color="blue" if "cr" in title.lower() else "green"),
-                    showlegend=True,
-                ),
-                row=idx,
-                col=1,
-            )
-
-            # Add 3-sigma bounds
-            df = df.with_columns(
-                [
-                    (pl.col(val_col) + 3.0 * pl.col(sigma_col)).alias(
-                        f"{title} +3-Sigma"
-                    ),
-                    (pl.col(val_col) - 3.0 * pl.col(sigma_col)).alias(
-                        f"{title} -3-Sigma"
-                    ),
-                ]
-            )
-            for bound in [f"{title} +3-Sigma", f"{title} -3-Sigma"]:
-                show_this_legend = not legend_added
-                fig.add_trace(
-                    go.Scatter(
-                        x=df["Epoch (UTC)"],
-                        y=df[bound],
-                        mode="lines",
-                        name="3-Sigma bounds",
-                        line=dict(color="black", dash="dash"),
-                        legendgroup="3-Sigma bounds",
-                        connectgaps=True,
-                        showlegend=show_this_legend,
-                    ),
-                    row=idx,
-                    col=1,
-                )
-                legend_added = True
-
-            fig.update_yaxes(title_text="Value (unitless)", row=idx, col=1)
-
-        fig.update_layout(
-            title_text=" ".join([x[0] for x in plots_to_make]), template=TEMPLATE
-        )
-        fig.update_xaxes(matches="x")
-        watermark(fig, path).show()
+    for plot in orbital_element_uncertainty(df, 3.0, path):
+        plot.show()
 
     if wstats:
-        for msr in msr_types:
-            px.scatter(
-                df,
-                x="Epoch (UTC)",
-                y=[f"Real observation: {msr}", f"Computed observation: {msr}"],
-            ).show()
+        gains_plot = kalman_gains(df, path)
+        if gains_plot is not None:
+            gains_plot.show()
 
-        gain_columns = [c for c in df.columns if "Gain" in c]
-        fs_ratio_columns = [c for c in df.columns if "Filter-smoother ratio" in c]
-        is_filter_run = len(df[gain_columns].drop_nulls()) > 0
-
-        # Plot the filter gains or filter-smoother ratios
-        if is_filter_run:
-            fig = px.scatter(df, x="Epoch (UTC)", y=gain_columns)
-            watermark(fig, path).show()
-        else:
-            fig = px.scatter(df, x="Epoch (UTC)", y=fs_ratio_columns)
-            watermark(fig, path).show()
+        fs_ratios_plot = filter_smoother_ratios(df, path)
+        if fs_ratios_plot is not None:
+            fs_ratios_plot.show()
 
 
 if __name__ == "__main__":
