+++ nyx-py/nyx_space/plots/plot_orbital_elements.py
@@ -1,7 +1,7 @@
 import click
-import plotly.graph_objs as go
 import polars as pl
-from plotly.subplots import make_subplots
+
+from nyx_space.plots.md import orbital_elements
 
 
 @click.command
@@ -12,46 +12,7 @@ def plot_orbit_elements(path: str):
     """
 
     df = pl.read_parquet(path)
-
-    df = df.with_columns(
-        pl.col("Epoch (UTC)").str.to_datetime("%Y-%m-%dT%H:%M:%S%.f")
-    ).sort("Epoch (UTC)", descending=False)
-
-    columns = [
-        "sma (km)",
-        "ecc",
-        "inc (deg)",
-        "raan (deg)",
-        "aop (deg)",
-        "ta (deg)",
-        "aol (deg)",
-        "tlong (deg)",
-    ]
-
-    fig = make_subplots(
-        rows=4,
-        cols=2,
-        subplot_titles=columns,
-        shared_xaxes=True,
-        vertical_spacing=0.1,
-    )
-
-    row_i = 0
-    col_i = 0
-
-    for col in columns:
-        name = col
-
-        fig.add_trace(
-            go.Scattergl(x=df["Epoch (UTC)"], y=df[col], name=name),
-            row=row_i + 1,
-            col=col_i + 1,
-        )
-        col_i = (col_i + 1) % 2
-        if col_i == 0:
-            row_i = (row_i + 1) % 4
-
-    fig.show()
+    orbital_elements(df).show()
 
 
 if __name__ == "__main__":
